Binary Search Tree Implementation
- Implemented the IsBalanced() method
- Massive refactor of the entire thing to handle a different way of checking whether the tree is balanced. Detailed below

--TreeNode--
Added height and weight fields. Got to thinking, and even though the weight and height of a tree _are_ attributes of the tree, they are only attributes which come from the nodes themselves. Consider the TreeNode as a trapeze artist - physically speaking, the minute they grasp the hands of another acrobat from their perch on the bar, the weight of the other acrobat becomes their weight as well: they take on the weight of the other acrobat, while the other acrobat only has their own weight.

The same is true for a tree: whether it is rooted in the ground or from a perch, it bears the weight of its children and their subtrees, and each child recursively bears the weight of its children and their subtrees, while the leaf nodes bear only their own weight (if at all; there is an argument to be made for having a node's weight and height initialized to 0 rather than 1, but this is semantics at this point). While the same is not necessarily true for the height - the person's height does not increase with the added mass, and so analogously neither would a node's with added weight - it is easy to follow the same principle, and much more convenient to simply update the height at each addition/removal with O(1) operations and reflect this with an O(1) call to the root, rather than do an O(log(n)) traversal to find the tree height.

--BinarySearchTree--
Massive overhaul of the Add and Remove functions to accomodate the above change to the TreeNode. Removed some redundant functions and the size attribute (which is now based on the weight of the root node).

- Add Function: Whenever a node is added, we make a log(n) traversal to wherever it is to be added (duplicates not allowed). From here we do a quick log(n) traversal back up the tree to update the weights (we can't do this before this point because we have to verify that the value CAN be added first), and if we have made a new level (i.e. added to a node with no children), then we propagate this fact through the recursive calls to update the heights at each node to reflect the max heights of their children.

- Remove Function: When a node is removed, we need to find it. Therefore we locate the node, do the removal, and propagate the changes up the tree so that all the relevant heights and weights are updated. When removing the node, we provide a reference to the data in case the user wants it.